requirements.txt
streamlit
openai
langchain
pymupdf
sentence-transformers
streamlit==1.27.0
python-dotenv==1.0.0
PyMuPDF==1.22.5
langchain==0.0.270
faiss-cpu==1.7.4
openai==1.3.0
pytesseract==0.3.10
Pillow==10.0.0
python-docx==0.8.11